app/src/test/java/com/fasten/executor_driver/presentation/services/ServicesViewStatePendingTest.java
package com.fasten.executor_driver.presentation.services;

import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.verifyNoMoreInteractions;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

@RunWith(MockitoJUnitRunner.class)
public class ServicesViewStatePendingTest {

  private ServicesViewStatePending viewState;

  @Mock
  private ServicesViewActions codeViewActions;

  @Before
  public void setUp() {
    viewState = new ServicesViewStatePending();
  }

  @Test
  public void testActions() {
    // Действие:
    viewState.apply(codeViewActions);

    // Результат:
    verify(codeViewActions).enableReadyButton(false);
    verify(codeViewActions).showServicesList(true);
    verify(codeViewActions).showServicesPending(true);
    verify(codeViewActions).showServicesListErrorMessage(false);
    verifyNoMoreInteractions(codeViewActions);
  }
}